# docs/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiEnvelope_NewsData_": {
        "additionalProperties": true,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NewsData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "ApiEnvelope[NewsData]",
        "type": "object"
      },
      "ApiEnvelope_NewsFactDetailData_": {
        "additionalProperties": true,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NewsFactDetailData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "ApiEnvelope[NewsFactDetailData]",
        "type": "object"
      },
      "ApiEnvelope_NewsObjectData_": {
        "additionalProperties": true,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NewsObjectData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "ApiEnvelope[NewsObjectData]",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NewsAgentAdmission": {
        "additionalProperties": true,
        "properties": {
          "basis": {
            "additionalProperties": true,
            "title": "Basis",
            "type": "object"
          },
          "eligible": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eligible"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "representative_news_item_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Representative News Item Id"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          }
        },
        "title": "NewsAgentAdmission",
        "type": "object"
      },
...
      "NewsData": {
        "additionalProperties": true,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NewsRow"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "title": "NewsData",
        "type": "object"
      },
      "NewsFactDetailData": {
        "additionalProperties": true,
        "properties": {
          "affected_targets_json": {
            "items": {},
            "title": "Affected Targets Json",
            "type": "array"
          },
          "canonical_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Url"
          },
          "claim": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claim"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "created_at_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At Ms"
          },
          "event_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Type"
          },
          "evidence_quote": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Quote"
          },
          "fact_candidate_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fact Candidate Id"
          },
          "headline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headline"
          },
          "news_item_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "News Item Id"
          },
          "realis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Realis"
          },
          "rejection_reasons_json": {
            "items": {
              "type": "string"
            },
            "title": "Rejection Reasons Json",
            "type": "array"
          },
          "source_domain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Domain"
          },
          "updated_at_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At Ms"
          },
          "validation_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validation Status"
          }
        },
        "title": "NewsFactDetailData",
        "type": "object"
      },
      "NewsFactLane": {
        "additionalProperties": true,
        "properties": {
          "affected_targets": {
            "items": {},
            "title": "Affected Targets",
            "type": "array"
          },
          "claim": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claim"
          },
          "event_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Type"
          },
          "realis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Realis"
          },
          "rejection_reasons": {
            "items": {
              "type": "string"
            },
            "title": "Rejection Reasons",
            "type": "array"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "title": "NewsFactLane",
        "type": "object"
      },
      "NewsMarketScope": {
        "additionalProperties": true,
        "properties": {
          "basis": {
            "additionalProperties": true,
            "title": "Basis",
            "type": "object"
          },
          "primary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "scope": {
            "items": {
              "type": "string"
            },
            "title": "Scope",
            "type": "array"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          }
        },
        "title": "NewsMarketScope",
        "type": "object"
      },
      "NewsObjectData": {
        "additionalProperties": true,
        "properties": {
          "agent_admission": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NewsAgentAdmission"
              },
              {
                "type": "null"
              }
            ]
          },
          "agent_admission_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Admission Reason"
          },
          "agent_admission_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Admission Status"
          },
          "agent_brief": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NewsAgentBrief"
              },
              {
                "type": "null"
              }
            ]
          },
          "agent_brief_computed_at_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Brief Computed At Ms"
          },
          "agent_brief_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Brief Status"
          },
          "agent_representative_news_item_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Representative News Item Id"
          },
          "agent_run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NewsAgentRunSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "agent_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Status"
          },
          "body_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body Text"
          },
          "canonical_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Url"
          },
          "computed_at_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Computed At Ms"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "content_class": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Class"
          },
          "content_classification": {
            "additionalProperties": true,
            "title": "Content Classification",
            "type": "object"
          },
          "content_tags": {
            "items": {
              "type": "string"
            },
            "title": "Content Tags",
            "type": "array"
          },
          "coverage_tags": {
            "items": {
              "type": "string"
            },
            "title": "Coverage Tags",
            "type": "array"
          },
          "duplicate_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duplicate Count"
          },
          "duplicate_observation_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duplicate Observation Count"
          },
          "entities": {
            "items": {},
            "title": "Entities",
            "type": "array"
          },
          "fact_candidate_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fact Candidate Id"
          },
          "fact_candidates": {
            "items": {
              "$ref": "#/components/schemas/NewsFactLane"
            },
            "title": "Fact Candidates",
            "type": "array"
          },
          "fact_lanes": {
            "items": {
              "$ref": "#/components/schemas/NewsFactLane"
            },
            "title": "Fact Lanes",
            "type": "array"
          },
          "fetch_run": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fetch Run"
          },
          "fetched_at_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fetched At Ms"
          },
          "headline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headline"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "latest_at_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest At Ms"
          },
          "lifecycle_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lifecycle Status"
          },
          "market_scope": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NewsMarketScope"
              },
              {
                "type": "null"
              }
            ]
          },
          "news_item_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "News Item Id"
          },
          "observation_edges": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Observation Edges",
            "type": "array"
          },
          "projection_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Projection Version"
          },
          "provider_item": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Item"
          },
          "provider_observations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Provider Observations",
            "type": "array"
          },
          "provider_rating": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NewsProviderRating"
              },
              {
                "type": "null"
              }
            ]
          },
          "provider_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Type"
          },
          "published_at_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At Ms"
          },
          "representative_news_item_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Representative News Item Id"
          },
          "row_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Row Id"
          },
          "signal": {
            "$ref": "#/components/schemas/NewsSignalEnvelope"
          },
          "source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NewsSourceSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "source_domain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Domain"
          },
          "source_domains": {
            "items": {
              "type": "string"
            },
            "title": "Source Domains",
            "type": "array"
          },
          "source_ids": {
            "items": {
              "type": "string"
            },
            "title": "Source Ids",
            "type": "array"
          },
          "source_quality_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Quality Status"
          },
          "source_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Role"
          },
          "story": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NewsStory"
              },
              {
                "type": "null"
              }
            ]
          },
          "story_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Story Key"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "token_impacts": {
            "items": {
              "$ref": "#/components/schemas/NewsTokenLane"
            },
            "title": "Token Impacts",
            "type": "array"
          },
          "token_lanes": {
            "items": {
              "$ref": "#/components/schemas/NewsTokenLane"
            },
            "title": "Token Lanes",
            "type": "array"
          },
          "token_mentions": {
            "items": {},
            "title": "Token Mentions",
            "type": "array"
          },
          "trust_tier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trust Tier"
          }
        },
        "title": "NewsObjectData",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Parallax",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/news": {
      "get": {
        "operationId": "list_news_api_news_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "default": "",
              "title": "Cursor",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "",
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "signal",
            "required": false,
            "schema": {
              "default": "",
              "title": "Signal",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_NewsData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List News",
        "tags": [
          "api"
        ]
      }
    },
    "/api/news/facts/{fact_candidate_id}": {
      "get": {
        "operationId": "get_news_fact_api_news_facts__fact_candidate_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "fact_candidate_id",
            "required": true,
            "schema": {
              "title": "Fact Candidate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_NewsFactDetailData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get News Fact",
        "tags": [
          "api"
        ]
      }
    },
    "/api/news/items/{news_item_id}": {
      "get": {
        "operationId": "get_news_item_api_news_items__news_item_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "news_item_id",
            "required": true,
            "schema": {
              "title": "News Item Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_NewsObjectData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get News Item",
        "tags": [
          "api"
        ]
      }
    },
...
  }
}